Handle referenced types.
I was wrong about what it meant before. Now that was easy... I want to
do a patch version bump (even though there was a version bump about 5
minutes ago) just to have some correctness around that. It's not a
feature that I use very much, but it is one that I _see_ used...